+++ admin/src/main/java/com/qiaomu/modules/propertycompany/controller/CommunityController.java
@@ -45,9 +45,14 @@ public class CommunityController extends AbstractController {
         return R.ok();
     }
 
+    /**
+     *  根据社区id获取社区信息
+     * @param id
+     * @return
+     */
     @ResponseBody
     @RequestMapping(value = "getCommunityById/{id}")
-    public R getCommunityById(@PathVariable("id") Long id) {
+    public R getCommunityById(@PathVariable("id") Integer id) {
         return R.ok("community", communityService.queryById(id));
     }
